video_core: Abstract backend agnostic rasterizer operations
